V1.5 新增图片下载功能&优化代码结构

+++ TaskFlow/pageCustomtaskFlow.h
@@ -27,7 +27,7 @@ private:
     void initRightUi();
     QWidget *rightWidget;
     QVBoxLayout *rightScrollWidgetLayout;
-    float m_leftratio = 0.45;
+    double m_leftratio = 0.45;
 
 protected:
     void resizeEvent(QResizeEvent *event) override;
